// source/frontend/src/models/api/generated/ApiGen_Concepts_LeaseAssociation.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { UtcIsoDate } from '@/models/api/UtcIsoDateTime';

import { ApiGen_Base_CodeType } from './ApiGen_Base_CodeType';
import { ApiGen_Concepts_LeaseStakeholder } from './ApiGen_Concepts_LeaseStakeholder';

// LINK: @backend/apimodels/Models/Concepts/Lease/LeaseAssociationModel.cs
export interface ApiGen_Concepts_LeaseAssociation {
  id: number;
  fileNumber: string | null;
  stakeholders: ApiGen_Concepts_LeaseStakeholder[] | null;
  leaseExpiryDate: UtcIsoDate | null;
  fileStatusTypeCode: ApiGen_Base_CodeType<string> | null;
  appCreateUserGuid: string | null;
  appCreateUserid: string | null;
}